re-organize the classes introduced as part of the networking stack refactor
