+++ auth0/src/main/java/com/auth0/android/authentication/AuthenticationAPIClient.java
@@ -36,14 +36,14 @@ import com.auth0.android.authentication.request.DatabaseConnectionRequest;
 import com.auth0.android.authentication.request.ProfileRequest;
 import com.auth0.android.authentication.request.SignUpRequest;
 import com.auth0.android.request.AuthenticationRequest;
+import com.auth0.android.request.DefaultClient;
+import com.auth0.android.request.ErrorAdapter;
+import com.auth0.android.request.JsonAdapter;
 import com.auth0.android.request.Request;
 import com.auth0.android.request.internal.BaseAuthenticationRequest;
+import com.auth0.android.request.internal.GsonAdapter;
 import com.auth0.android.request.internal.GsonProvider;
 import com.auth0.android.request.internal.RequestFactory;
-import com.auth0.android.request.kt.DefaultClient;
-import com.auth0.android.request.kt.ErrorAdapter;
-import com.auth0.android.request.kt.GsonAdapter;
-import com.auth0.android.request.kt.JsonAdapter;
 import com.auth0.android.result.Credentials;
 import com.auth0.android.result.DatabaseUser;
 import com.auth0.android.result.UserProfile;
