+++ auth0/src/main/java/com/auth0/android/management/UsersAPIClient.java
@@ -33,14 +33,14 @@ import androidx.annotation.VisibleForTesting;
 import com.auth0.android.Auth0;
 import com.auth0.android.Auth0Exception;
 import com.auth0.android.authentication.ParameterBuilder;
+import com.auth0.android.request.DefaultClient;
+import com.auth0.android.request.ErrorAdapter;
+import com.auth0.android.request.JsonAdapter;
+import com.auth0.android.request.NetworkingClient;
 import com.auth0.android.request.Request;
+import com.auth0.android.request.internal.GsonAdapter;
 import com.auth0.android.request.internal.GsonProvider;
 import com.auth0.android.request.internal.RequestFactory;
-import com.auth0.android.request.kt.DefaultClient;
-import com.auth0.android.request.kt.ErrorAdapter;
-import com.auth0.android.request.kt.GsonAdapter;
-import com.auth0.android.request.kt.JsonAdapter;
-import com.auth0.android.request.kt.NetworkingClient;
 import com.auth0.android.result.UserIdentity;
 import com.auth0.android.result.UserProfile;
 import com.auth0.android.util.Auth0UserAgent;
