+++ auth0/src/main/java/com/auth0/android/request/HttpMethod.kt
@@ -1,4 +1,4 @@
-package com.auth0.android.request.kt
+package com.auth0.android.request
 
 public sealed class HttpMethod {
