+++ auth0/src/main/java/com/auth0/android/request/NetworkingClient.kt
@@ -1,4 +1,4 @@
-package com.auth0.android.request.kt
+package com.auth0.android.request
 
 import java.io.IOException
