+++ auth0/src/main/java/com/auth0/android/request/RequestOptions.kt
@@ -1,4 +1,4 @@
-package com.auth0.android.request.kt
+package com.auth0.android.request
 
 /**
  * Holder for the information required to configure a request
