+++ auth0/src/main/java/com/auth0/android/request/internal/ThreadUtils.kt
@@ -1,4 +1,4 @@
-package com.auth0.android.request.kt
+package com.auth0.android.request.internal
 
 import android.os.Handler
 import android.os.Looper
@@ -6,6 +6,9 @@ import androidx.core.os.HandlerCompat
 import java.util.concurrent.ExecutorService
 import java.util.concurrent.Executors
 
+/**
+ * Utilities to quickly switch threads.
+ */
 internal object ThreadUtils {
     private const val MAX_CONCURRENT_THREADS = 4
