+++ auth0/src/test/java/com/auth0/android/request/internal/BaseRequestTest.java
@@ -26,13 +26,12 @@ package com.auth0.android.request.internal;
 
 
 import com.auth0.android.Auth0Exception;
-import com.auth0.android.request.kt.ErrorAdapter;
-import com.auth0.android.request.kt.GsonAdapter;
-import com.auth0.android.request.kt.HttpMethod;
-import com.auth0.android.request.kt.JsonAdapter;
-import com.auth0.android.request.kt.NetworkingClient;
-import com.auth0.android.request.kt.RequestOptions;
-import com.auth0.android.request.kt.ServerResponse;
+import com.auth0.android.request.ErrorAdapter;
+import com.auth0.android.request.HttpMethod;
+import com.auth0.android.request.JsonAdapter;
+import com.auth0.android.request.NetworkingClient;
+import com.auth0.android.request.RequestOptions;
+import com.auth0.android.request.ServerResponse;
 import com.google.gson.Gson;
 
 import org.hamcrest.collection.IsMapContaining;
