+++ auth0/src/test/java/com/auth0/android/request/internal/RequestFactoryTest.java
@@ -2,11 +2,11 @@ package com.auth0.android.request.internal;
 
 
 import com.auth0.android.Auth0Exception;
+import com.auth0.android.request.ErrorAdapter;
+import com.auth0.android.request.HttpMethod;
+import com.auth0.android.request.JsonAdapter;
+import com.auth0.android.request.NetworkingClient;
 import com.auth0.android.request.Request;
-import com.auth0.android.request.kt.ErrorAdapter;
-import com.auth0.android.request.kt.HttpMethod;
-import com.auth0.android.request.kt.JsonAdapter;
-import com.auth0.android.request.kt.NetworkingClient;
 
 import org.junit.Before;
 import org.junit.Test;
